backend/requirements.txt
fastapi==0.115.6
GitPython==3.1.44
qdrant-client==1.13.0
transformers==4.48.1
tree-sitter==0.20.1
python-dotenv==1.0.1
openai==1.59.9
langchain-text-splitters==0.3.5
deepeval==2.1.9
nltk==3.9.1
python-multipart==0.0.20
uvicorn==0.34.0
aioboto3==13.4.0